refactor code_generator
