+++ code_generator/base.py
@@ -1,8 +1,9 @@
-import jinja2
 import os
+import jinja2
 import json
 import csv
 from utils import unique,flatten
+
 from mutations import MutationFasta
 
 class CodeGenerator(object):
@@ -11,7 +12,36 @@ class CodeGenerator(object):
         self.template_dir = os.path.join(os.path.dirname(__file__), 'templates')
         self.jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(self.template_dir),
                                autoescape = False,
-                               extensions=['jinja2.ext.with_'])        
+                               extensions=['jinja2.ext.with_'])  
+        self.gene_enum_to_drug_name = self.load_gene_enum_to_drug_name()
+        self.virulence_genes = unique(self.load_virulence_genes())
+        self.template_dir = os.path.join(os.path.dirname(__file__), 'templates/' )
+        self.render_dir = os.path.join(os.path.dirname(__file__), 'rendered/' )
+
+        self.jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(self.template_dir),
+                               autoescape = False,
+                               extensions=['jinja2.ext.with_'])  
+
+    @property 
+    def genes(self):
+        return unique(self.gene_enum_to_drug_name.keys() + self.virulence_genes)
+
+    @property 
+    def drugs(self):
+        raise NotImplementedError("Implemented in child")
+
+    @property 
+    def mutation_induced_drug_names(self):
+        raise NotImplementedError("Implemented in child")
+
+    @property 
+    def drug_names(self):
+        drug_names = unique(self.gene_induced_drug_names + self.mutation_induced_drug_names) 
+        return drug_names                        
+
+    @property 
+    def gene_induced_drug_names(self):
+        return unique(flatten(self.gene_enum_to_drug_name.values()))
        
     def render_str(self, template, **params):
         try:
@@ -32,55 +62,20 @@ class CodeGenerator(object):
             row = reader.next()
         return row
 
-
-
-class StaphCodeGenerator(CodeGenerator):
-
-    def __init__(self):
-        self.species = "staph"
-        self.gene_enum_to_drug_name = self.load_gene_enum_to_drug_name()
-        self.virulence_genes = unique(self.load_virulence_genes())
-        self.template_dir = os.path.join(os.path.dirname(__file__), 'templates/' )
-        self.render_dir = os.path.join(os.path.dirname(__file__), 'rendered/' )
-
-        self.jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(self.template_dir),
-                               autoescape = False,
-                               extensions=['jinja2.ext.with_'])
-    @property 
-    def genes(self):
-        return unique(self.gene_enum_to_drug_name.keys() + self.virulence_genes)
-
-    @property 
-    def drugs(self):
-        return [StaphDrugCodeGenerator(drug) for drug in self.drug_names ]
-
-    @property 
-    def drug_names(self):
-        drug_names = unique(self.gene_induced_drug_names + self.mutation_induced_drug_names) 
-        drug_names.insert(0, drug_names.pop(drug_names.index("Erythromycin")))
-        return drug_names                        
-
-    @property 
-    def gene_induced_drug_names(self):
-        return unique(flatten(self.gene_enum_to_drug_name.values()))
-
-    @property 
-    def mutation_induced_drug_names(self):
-        return ['Rifampicin','Ciprofloxacin']
-
     def render_antibiotics_src(self):
-        return self.render_str('src/predictor/staph/antibiotics.c')
+        return self.render_str('src/predictor/%s/antibiotics.c' % self.species)        
 
     def render_antibiotics_include(self):
-        return self.render_str('include/predictor/staph/antibiotics.h')        
+        return self.render_str('include/predictor/%s/antibiotics.h' % self.species)   
+
 
     def render_and_write_antibiotics(self):
         st = self.render_antibiotics_src()
-        with open(os.path.join(self.render_dir,'src/predictor/staph/antibiotics.c'),'w') as outfile:
+        with open(os.path.join(self.render_dir,'src/predictor/%s/antibiotics.c' % self.species),'w') as outfile:
             outfile.write(st)
         st = self.render_antibiotics_include()
-        with open(os.path.join(self.render_dir,'include/predictor/staph/antibiotics.h'),'w') as outfile:
-            outfile.write(st) 
+        with open(os.path.join(self.render_dir,'include/predictor/%s/antibiotics.h' % self.species),'w') as outfile:
+            outfile.write(st)         
 
     def render_known_mutations_src(self):
         return self.render_str('src/predictor/core/known_mutations.c')
@@ -96,6 +91,7 @@ class StaphCodeGenerator(CodeGenerator):
         with open(os.path.join(self.render_dir,'include/predictor/core/known_mutations.h'),'w') as outfile:
             outfile.write(st) 
 
+
     def render_gene_presence_src(self):
         return self.render_str('src/predictor/core/gene_presence.c')
 
@@ -110,13 +106,15 @@ class StaphCodeGenerator(CodeGenerator):
         with open(os.path.join(self.render_dir,'include/predictor/core/gene_presence.h'),'w') as outfile:
             outfile.write(st) 
 
+
     def render_main_src(self):
-        return self.render_str('src/predictor/staph/main.c')
+        return self.render_str('src/predictor/%s/main.c' % self.species )
 
     def render_and_write_main(self):
         st = self.render_main_src()
-        with open(os.path.join(self.render_dir,'src/predictor/staph/main.c'),'w') as outfile:
-            outfile.write(st)    
+        with open(os.path.join(self.render_dir,'src/predictor/%s/main.c' % self.species),'w') as outfile:
+            outfile.write(st)   
+
 
     def render_and_write_all(self):
         self.render_and_write_antibiotics() 
@@ -137,35 +135,25 @@ class StaphCodeGenerator(CodeGenerator):
         mut_genes = []
         for drug in self.drugs:
             mut_genes.extend(drug.mut_gene_list)
-        return mut_genes        
-
-
-
+        return mut_genes      
 
 class DrugCodeGenerator(CodeGenerator):
 
-    def __init__(self,name):
+    def __init__(self, name):
         self.name = name
         self.gene_enum_to_drug_name = self.load_gene_enum_to_drug_name()
-        self.mutation_fasta = MutationFasta('../data/staph/antibiotics/%s' % self.name.lower() )
+        self.mutation_fasta = MutationFasta('../data/%s/antibiotics/%s' % (self.species, self.name.lower() )  )
 
     def __str__(self):
         return self.name   
 
     @property
     def genes_resistance_induced_by(self):
-        genes = []
-        for gene,drugs in self.gene_enum_to_drug_name.iteritems():
-            if self.name in drugs:
-                genes.append(StaphGene(gene))
-        return genes 
+        raise NotImplementedError("Implemented in child")          
 
     @property 
     def has_epistatic_muts(self):
-        if self.name in ['FusidicAcid','Rifampicin']:
-            return True
-        else:
-            return False
+        raise NotImplementedError("Implemented in child")          
 
     @property 
     def epistaic_file_path(self):
@@ -193,14 +181,8 @@ class DrugCodeGenerator(CodeGenerator):
 
     @property 
     def num_genes(self):
-        return len(self.genes_resistance_induced_by)                   
+        return len(self.genes_resistance_induced_by)      
 
-class StaphDrugCodeGenerator(DrugCodeGenerator):
-    
-    def __init__(self,name):
-        self.species = "staph"
-        super(StaphDrugCodeGenerator, self).__init__(name)            
-        
 
 class GeneCodeGenerator(CodeGenerator):
 
@@ -209,20 +191,4 @@ class GeneCodeGenerator(CodeGenerator):
         self.gene_enum_to_drug_name = self.load_gene_enum_to_drug_name()
 
     def __str__(self):
-        return self.name     
-
-        
-class StaphGene(GeneCodeGenerator):
-    
-    def __init__(self,name):
-        self.species = "staph" 
-        super(StaphGene, self).__init__(name=name)                    
-                               
-
-
-cg = StaphCodeGenerator()
-cg.render_and_write_all()
-# print cg.render_map_gene_to_drug_resistance()
-# print cg.render_antibiotics()
-# print cg.is_drug_susceptible(drug="Gentamicin")
-# print cg.is_drug_susceptible(drug="FusidicAcid")
+        return self.name
